Fix: migrate 파일 내 각 모델의 id type int로 수정
Bigint => Integer 로 데이터 타입 수정.
추후에 변경해야할 상황이 생기면 그때 Bigint로 변경

+++ api-server/migrations/20191113062914-create-video.js
@@ -5,7 +5,7 @@ module.exports = {
         allowNull: true,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.BIGINT(11),
+        type: Sequelize.INTEGER,
       },
       name: {
         type: Sequelize.STRING,
